Updated the active speaker detection
