+++ template/src/components/useVideoCall.tsx
@@ -19,6 +19,8 @@ import {useRtc, useMeetingInfo} from 'customization-api';
 import SDKEvents from '../utils/SdkEvents';
 import DeviceContext from './DeviceContext';
 import useSetName from '../utils/useSetName';
+import useFindActiveSpeaker from '../utils/useFindActiveSpeaker';
+import {UidType} from 'customization-api';
 
 export interface VideoCallContextInterface {
   showInvitePopup: boolean;
@@ -27,6 +29,7 @@ export interface VideoCallContextInterface {
   setShowStopRecordingPopup: React.Dispatch<SetStateAction<boolean>>;
   showLayoutOption: boolean;
   setShowLayoutOption: React.Dispatch<SetStateAction<boolean>>;
+  activeSpeaker: UidType;
 }
 
 const VideoCallContext = React.createContext<VideoCallContextInterface>({
@@ -36,6 +39,7 @@ const VideoCallContext = React.createContext<VideoCallContextInterface>({
   setShowStopRecordingPopup: () => {},
   showLayoutOption: false,
   setShowLayoutOption: () => {},
+  activeSpeaker: 0,
 });
 
 interface VideoCallProviderProps {
@@ -49,6 +53,7 @@ const VideoCallProvider = (props: VideoCallProviderProps) => {
   const meetingInfo = useMeetingInfo();
   const {deviceList} = useContext(DeviceContext);
   const setUsername = useSetName();
+  const activeSpeaker = useFindActiveSpeaker();
 
   useEffect(() => {
     if (join.initialized && join.phrase) {
@@ -76,6 +81,7 @@ const VideoCallProvider = (props: VideoCallProviderProps) => {
         setShowStopRecordingPopup,
         showLayoutOption,
         setShowLayoutOption,
+        activeSpeaker,
       }}>
       <StopRecordingPopup />
       <InvitePopup />
