+++ template/src/utils/useIsLocalUserSpeaking.native.ts
@@ -1,0 +1,4 @@
+const useIsLocalUserSpeaking = () => {
+  return () => {};
+};
+export default useIsLocalUserSpeaking;
