--- template/src/utils/useIsSpeaking.native.tsx
@@ -1,4 +1,0 @@
-const useIsSpeaking = () => {
-  return () => {};
-};
-export default useIsSpeaking;
